Move repo store to gitserver package.
