+++ internal/codeintel/stores/gitserver/repo.go
@@ -1,4 +1,4 @@
-package dbstore
+package gitserver
 
 import (
 	"context"
@@ -6,23 +6,27 @@ import (
 
 	"github.com/keegancsmith/sqlf"
 	"github.com/lib/pq"
-	"github.com/opentracing/opentracing-go/log"
 
+	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
-	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
+
+type store struct {
+	*basestore.Store
+}
+
+func newWithDB(db database.DB) *store {
+	return &store{
+		Store: basestore.NewWithHandle(db.Handle()),
+	}
+}
 
 // ErrUnknownRepository occurs when a repository does not exist.
 var ErrUnknownRepository = errors.New("unknown repository")
 
 // RepoName returns the name for the repo with the given identifier.
-func (s *Store) RepoName(ctx context.Context, repositoryID int) (_ string, err error) {
-	ctx, _, endObservation := s.operations.repoName.With(ctx, &err, observation.Args{LogFields: []log.Field{
-		log.Int("repositoryID", repositoryID),
-	}})
-	defer endObservation(1, observation.Args{})
-
+func (s *store) RepoName(ctx context.Context, repositoryID int) (_ string, err error) {
 	name, exists, err := basestore.ScanFirstString(s.Store.Query(ctx, sqlf.Sprintf(repoNameQuery, repositoryID)))
 	if err != nil {
 		return "", err
@@ -36,6 +40,16 @@ func (s *Store) RepoName(ctx context.Context, repositoryID int) (_ string, err e
 const repoNameQuery = `
 -- source: internal/codeintel/stores/dbstore/repos.go:RepoName
 SELECT name FROM repo WHERE id = %s
+`
+
+// RepoNames returns a map from repository id to names.
+func (s *store) RepoNames(ctx context.Context, repositoryIDs ...int) (_ map[int]string, err error) {
+	return scanRepoNames(s.Store.Query(ctx, sqlf.Sprintf(repoNamesQuery, pq.Array(repositoryIDs))))
+}
+
+const repoNamesQuery = `
+-- source: internal/codeintel/stores/dbstore/repos.go:RepoNames
+SELECT id, name FROM repo WHERE id = ANY(%s)
 `
 
 func scanRepoNames(rows *sql.Rows, queryErr error) (_ map[int]string, err error) {
@@ -60,18 +74,3 @@ func scanRepoNames(rows *sql.Rows, queryErr error) (_ map[int]string, err error)
 
 	return names, nil
 }
-
-// RepoNames returns a map from repository id to names.
-func (s *Store) RepoNames(ctx context.Context, repositoryIDs ...int) (_ map[int]string, err error) {
-	ctx, _, endObservation := s.operations.repoName.With(ctx, &err, observation.Args{LogFields: []log.Field{
-		log.Int("numRepositories", len(repositoryIDs)),
-	}})
-	defer endObservation(1, observation.Args{})
-
-	return scanRepoNames(s.Store.Query(ctx, sqlf.Sprintf(repoNamesQuery, pq.Array(repositoryIDs))))
-}
-
-const repoNamesQuery = `
--- source: internal/codeintel/stores/dbstore/repos.go:RepoNames
-SELECT id, name FROM repo WHERE id = ANY(%s)
-`
